requirements.txt
pip~=21.3.1
setuptools~=58.3.0
python-dotenv==0.19.1
pyTelegramBotAPI==4.1.1
requests==2.26.0
urllib3==1.26.7
idna==3.3
python-telegram-bot-calendar==1.0.5